requirements.txt
python-dotenv
discord.py[voice]
yt-dlp
PyNaCl
